+++ tests/ui/const-generics/issue-93647.rs
@@ -1,6 +1,7 @@
 struct X<const N: usize = {
     (||1usize)()
     //~^ ERROR cannot call non-const closure
+    //~| ERROR the trait bound
 }>;
 
 fn main() {}
