+++ tests/ui/consts/issue-28113.rs
@@ -3,6 +3,7 @@
 const X: u8 =
     || -> u8 { 5 }()
     //~^ ERROR cannot call non-const closure
+    //~| ERROR the trait bound
 ;
 
 fn main() {}
