+++ tests/ui/consts/issue-56164.rs
@@ -1,5 +1,6 @@
 const fn foo() { (||{})() }
 //~^ ERROR cannot call non-const closure
+//~| ERROR the trait bound
 
 const fn bad(input: fn()) {
     input()
