+++ tests/ui/consts/issue-68542-closure-in-array-len.rs
@@ -4,6 +4,7 @@
 
 struct Bug {
     a: [(); (|| { 0 })()] //~ ERROR cannot call non-const closure
+    //~^ ERROR the trait bound
 }
 
 fn main() {}
