+++ tests/ui/never_type/issue-52443.rs
@@ -11,4 +11,5 @@ fn main() {
     //~| ERROR cannot convert
     //~| ERROR mutable references
     //~| ERROR cannot call
+    //~| ERROR the trait bound
 }
