+++ tests/ui/rfc-2632-const-trait-impl/const_derives/derive-const-non-const-type.rs
@@ -9,5 +9,6 @@ impl Default for A {
 #[derive_const(Default)]
 pub struct S(A);
 //~^ cannot call non-const fn
+//~| the trait bound
 
 fn main() {}
